separated dagger module for db
